+++ measure.h
@@ -10,5 +10,5 @@ void MeasureInit(void);
  * @param[out] the current (amperes)
  * @param[out] the voltage (volts)
  */
-void MeasureGet(uint32_t *time, float *current, float *voltage);
+void MeasureGet(unsigned long *time, float *current, float *voltage);
